@@ -9,7 +9,7 @@ authors = [{name = "Timothée Mazzucotelli", email = "dev@pawamoy.fr"}]
 license = "ISC"
 license-files = ["LICENSE"]
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 keywords = []
 dynamic = ["version"]
 classifiers = [
@@ -18,7 +18,6 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
@@ -49,7 +48,7 @@ getter = "scripts.get_version:get_version"
 
 [tool.pdm.build]
 # Include as much as possible in the source distribution, to help redistributors.
-excludes = ["**/.pytest_cache", "**/.mypy_cache"]
+excludes = ["**/.pytest_cache"]
 source-includes = [
     "config",
     "docs",
@@ -57,7 +56,7 @@ source-includes = [
     "share",
     "tests",
     "duties.py",
-    "mkdocs.yml",
+    "zensical.toml",
     "*.md",
     "LICENSE",
 ]
@@ -72,35 +71,30 @@ data = [
 
 [dependency-groups]
 maintain = [
-    "build>=1.2",
     "git-changelog>=2.5",
     "twine>=5.1",
     "yore>=0.3.3",
 ]
 ci = [
     "duty>=1.6",
-    "ruff>=0.4",
+    "griffe>=2.0",
     "pytest>=8.2",
     "pytest-cov>=5.0",
     "pytest-randomly>=3.15",
     "pytest-xdist>=3.6",
-    "mypy>=1.10",
+    "ruff>=0.4",
+    "ty>=0.0.14",
     "types-markdown>=3.6",
     "types-pyyaml>=6.0",
 ]
  docs = [
+    "ghp-import>=2.1",
     "markdown-callouts>=0.4",
     "markdown-exec>=1.8",
-    "mkdocs>=1.6",
-    "mkdocs-coverage>=1.0",
-    "mkdocs-git-revision-date-localized-plugin>=1.2",
-    "mkdocs-llmstxt>=0.2",
-    "mkdocs-material>=9.5",
-    "mkdocs-minify-plugin>=0.8",
-    "mkdocs-section-index>=0.3",
     "mkdocstrings[python]>=0.29",
     # YORE: EOL 3.10: Remove line.
     "tomli>=2.0; python_version < '3.11'",
+    "zensical>=0.0.21",
 ]
 
 [tool.uv]
